placement: Fix best cost estimation
https://code.google.com/p/irqbalance/issues/detail?id=48
Notes that our irq load estimates are off, because we truncate best cost at a 32
bit integer, even though we use a 64 bit uint to store it.  Fix the initial
best_cost to be ULLONG_MAX

+++ placement.c
@@ -115,7 +115,7 @@ static void find_best_object_for_irq(struct irq_info *info, void *data)
 	place.info = info;
 	place.best = NULL;
 	place.least_irqs = NULL;
-	place.best_cost = INT_MAX;
+	place.best_cost = ULLONG_MAX;
 
 	for_each_object(d->children, find_best_object, &place);
